+++ .gitignore
@@ -34,3 +34,4 @@ landcover/
 
 *.mbtiles
 *.pbf
+.aider*
