+++ cache.ts
@@ -1,0 +1,8 @@
+import Cache from 'file-system-cache'
+const month = 60 * 60 * 24 * 30
+const cache = Cache.default({
+  basePath: './.cache', // (optional) Path where cache files are stored (default).
+  ttl: month, // (optional) A time-to-live (in secs) on how long an item remains cached.
+})
+
+export default cache
